refactor(liferay-cli): move sample configuration generation to a facet
