+++ projects/js-toolkit/packages/liferay-cli/src/new/target-liferay/index.ts
@@ -16,6 +16,7 @@ import facetPlainJs from '../facet-plain-js';
 import facetPortlet from '../facet-portlet';
 import facetProject from '../facet-project';
 import facetReact from '../facet-react';
+import facetSampleConfiguration from '../facet-sample-configuration';
 import facetSampleStyles from '../facet-sample-styles';
 import prompt from '../util/prompt';
 
@@ -64,6 +65,7 @@ const target: Target = {
 
 		if (frameworkFacet) {
 			options = await frameworkFacet.prompt(useDefaults, options);
+			options = await facetSampleConfiguration.prompt(true, options);
 			options = await facetSampleStyles.prompt(true, options);
 			options = await facetBuildable.prompt(true, options);
 		}
@@ -83,6 +85,7 @@ const target: Target = {
 
 		if (frameworkFacet) {
 			await frameworkFacet.render(options);
+			await facetSampleConfiguration.render(options);
 			await facetSampleStyles.render(options);
 			await facetBuildable.render(options);
 		}
